+++ src/convtools/contrib/models/validators/__init__.py
@@ -1,3 +1,14 @@
 """Exposes validators - objects which only check the data and populate error
 objects"""
-from .validators import Custom, Decimal, Gt, Gte, Length, Lt, Lte, Regex, Type
+from .validators import (
+    Custom,
+    Decimal,
+    Enum,
+    Gt,
+    Gte,
+    Length,
+    Lt,
+    Lte,
+    Regex,
+    Type,
+)